# 修正woc openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "修正woc",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/academy/judge": {
      "post": {
        "summary": "创建评委",
        "deprecated": false,
        "description": "",
        "tags": [
          "academy"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userCode": {
                    "type": "string"
                  },
                  "comId": {
                    "type": "integer"
                  }
                },
                "required": [
                  "userCode",
                  "comId"
                ],
                "x-apifox-orders": [
                  "userCode",
                  "comId"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "x-apifox-orders": [],
                      "x-apifox-ignore-properties": []
                    }
                  },
                  "required": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/academy",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845404-run"
      },
      "delete": {
        "summary": "删除评委",
        "deprecated": false,
        "description": "",
        "tags": [
          "academy"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userCode": {
                    "type": "integer"
                  }
                },
                "required": [
                  "userCode"
                ],
                "x-apifox-orders": [
                  "userCode"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/academy",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845405-run"
      },
      "get": {
        "summary": "评委列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "academy"
        ],
        "parameters": [
          {
            "name": "comId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "x-apifox-orders": [],
                      "x-apifox-ignore-properties": []
                    }
                  },
                  "required": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/academy",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845406-run"
      }
    },
    "/judge": {
      "post": {
        "summary": "评分",
        "deprecated": false,
        "description": "",
        "tags": [
          "judge"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Score"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/judge",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845407-run"
      },
      "patch": {
        "summary": "修改评分",
        "deprecated": false,
        "description": "",
        "tags": [
          "judge"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Score"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/judge",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845408-run"
      }
    },
    "/admin/competition": {
      "post": {
        "summary": "创建比赛（必做）",
        "deprecated": false,
        "description": "##### 超级管理员创建比赛\n\n",
        "tags": [
          "admin"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "minTeamMembers": {
                    "type": "integer"
                  },
                  "maxTeamMembers": {
                    "type": "integer"
                  },
                  "startTime": {
                    "type": "string"
                  },
                  "endTime": {
                    "type": "string"
                  },
                  "reviewBeginTime": {
                    "type": "string"
                  },
                  "reviewEndTime": {
                    "type": "string"
                  }
                },
                "required": [
                  "minTeamMembers",
                  "description",
                  "name",
                  "reviewEndTime",
                  "reviewBeginTime",
                  "endTime",
                  "startTime",
                  "maxTeamMembers"
                ],
                "x-apifox-orders": [
                  "name",
                  "description",
                  "minTeamMembers",
                  "maxTeamMembers",
                  "startTime",
                  "endTime",
                  "reviewBeginTime",
                  "reviewEndTime"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "minTeamMembers": {
                          "type": "integer"
                        },
                        "maxTeamMembers": {
                          "type": "integer"
                        },
                        "workCode": {
                          "type": "string"
                        },
                        "startTime": {
                          "type": "string"
                        },
                        "endTime": {
                          "type": "string"
                        },
                        "reviewBeginTime": {
                          "type": "string"
                        },
                        "reviewEndTime": {
                          "type": "string"
                        },
                        "createTime": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "description",
                        "minTeamMembers",
                        "maxTeamMembers",
                        "workCode",
                        "startTime",
                        "endTime",
                        "reviewBeginTime",
                        "reviewEndTime",
                        "createTime"
                      ],
                      "x-apifox-orders": [
                        "id",
                        "name",
                        "description",
                        "minTeamMembers",
                        "maxTeamMembers",
                        "workCode",
                        "startTime",
                        "endTime",
                        "reviewBeginTime",
                        "reviewEndTime",
                        "createTime"
                      ],
                      "x-apifox-ignore-properties": []
                    }
                  },
                  "required": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/admin",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845409-run"
      }
    },
    "/admin/academy": {
      "post": {
        "summary": "创建二级管理员",
        "deprecated": false,
        "description": "",
        "tags": [
          "admin"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userCode": {
                    "type": "string"
                  },
                  "academyId": {
                    "type": "integer"
                  },
                  "comId": {
                    "type": "integer"
                  }
                },
                "required": [
                  "userCode",
                  "academyId",
                  "comId"
                ],
                "x-apifox-orders": [
                  "userCode",
                  "academyId",
                  "comId"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "x-apifox-orders": [],
                      "x-apifox-ignore-properties": []
                    }
                  },
                  "required": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/admin",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845410-run"
      },
      "delete": {
        "summary": "删除二级管理员",
        "deprecated": false,
        "description": "",
        "tags": [
          "admin"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userCode": {
                    "type": "string"
                  }
                },
                "required": [
                  "userCode"
                ],
                "x-apifox-orders": [
                  "userCode"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/admin",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845411-run"
      }
    },
    "/admin/judge": {
      "post": {
        "summary": "创建评委",
        "deprecated": false,
        "description": "",
        "tags": [
          "admin"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userCode": {
                    "type": "string"
                  },
                  "academyId": {
                    "type": "integer"
                  },
                  "comId": {
                    "type": "integer"
                  }
                },
                "required": [
                  "userCode",
                  "comId",
                  "academyId"
                ],
                "x-apifox-orders": [
                  "userCode",
                  "academyId",
                  "comId"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "x-apifox-orders": [],
                      "x-apifox-ignore-properties": []
                    }
                  },
                  "required": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/admin",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845412-run"
      },
      "delete": {
        "summary": "删除评委",
        "deprecated": false,
        "description": "",
        "tags": [
          "admin"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userCode": {
                    "type": "integer"
                  }
                },
                "required": [
                  "userCode"
                ],
                "x-apifox-orders": [
                  "userCode"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/admin",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845413-run"
      },
      "get": {
        "summary": "评委列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "name": "comId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "x-apifox-orders": [],
                      "x-apifox-ignore-properties": []
                    }
                  },
                  "required": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/admin",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845414-run"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GlobalResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "errCode": {
            "type": "integer"
          },
          "errMsg": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {},
            "x-apifox-orders": [],
            "x-apifox-ignore-properties": []
          }
        },
        "required": [
          "success",
          "data",
          "errMsg",
          "errCode"
        ],
        "x-apifox-orders": [
          "success",
          "errCode",
          "errMsg",
          "data"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
      "Score": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "judgeId": {
            "type": "string"
          },
          "workCode": {
            "type": "string"
          },
          "teamId": {
            "type": "integer"
          },
          "comId": {
            "type": "integer"
          },
          "score": {
            "type": "integer"
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "judgeId",
          "workCode",
          "teamId",
          "comId",
          "score",
          "comment"
        ],
        "x-apifox-orders": [
          "id",
          "judgeId",
          "workCode",
          "teamId",
          "comId",
          "score",
          "comment"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": []
}